Clarified the resourceFilters usage
Not 100% sure that I got this correct. Please review and fix.

I also see in the source code that the entries can also be just `[<Kind>]` or `[<Kind>,<Namespace>]` (i.e. omitting trailing parts instead of using `*`) but I'm not sure if that should be used so I did not mention it.