Don't check for Prom Operator apiVersion
`.Capabilities.APIVersions.Has` function has limitations when running with
`helm template`, which is common step in multiple CD tools. In order to
properly resolve `Capabilities.APIVersions` `helm template` has to run
with `--validate` option and connect to cluster that has Prom Operator
CRDs installed.

As this template is opt-in and user has to set value to enable this,
apiVersion check doesn't provide much value and can be removed.

Signed-off-by: Ihor Urazov <iurazov@healthjoy.com>