fix: only evaluate tests for matching engine responses
We've got to make sure we don't evaluate tests against EngineReponses
which have nothing to do with the current test. Failing to do so will
lead to wrongfully interpreting test results.

For example, we might look at the first EngineResponse, but compare
the generatedResource to that of the last test definition. This
would not line up and set the wrong test result.

Signed-off-by: Peter Bücker <peter.buecker@form3.tech>