Update dashboard.json
Current query aggregates data across multiple services and pods, which results in duplicates. e.g. 24 validate rules across 3 pods will show as 72 rules. They have to be consolidated by rule-names. 

Also, auto-gen rules are auto-generated and are not actual reflection of rules in the policy. So, they are being removed by default. Users can feel free to remove the autogen filter in their queries if they would like to the rule to be included.

Signed-off-by: Anubhav Sharma <40705688+anubhav888@users.noreply.github.com>