feat: add watch failure check
The WatchErrorHandler is called whenever ListAndWatch drops the connection with an error. By default, the error is reported in logs but the end user of resource cache will not know what error has been reported. This custom WatchErrorHandler will store the error received.

Since the watch handler is only called when there is an error, we have to clear the error after sometime ourselves. This handler uses an ephemeral store which will clear the data after sometime. We have set the clear time at slightly above the informer resync period.

Signed-off-by: Vishal Choudhary <vishal.choudhary@nirmata.com>