chore: bump cosign to v2.2.4 (#10039)
Cosign v2.2.4 has fixes for two advisories for denial of service vulnerabilities - GHSA-95pr-fxf5-86gv and GHSA-88jx-383q-w4qc. This PR bumps to the latest version.

The upstream version of cosign uses prometheus/common version v0.51.1.
There was a breaking change in v0.48.0 of prometheus/common and k8s.io/component-base@v1.29.3 uses a lower version (v0.44.0) which is not compatible with v0.51.1.

To fix this, I have created a new tag in our folk of cosign v2.2.4-deps-fix. Which takes the latest cosign v.2.2.4 and adds a commit to downgrade prometheus common to v0.47.0. Here is the commit history: kyverno/cosign@v2.2.4-deps-fix (commits)

We have to use this folk until k8s.io/component-base releases a new version with the bump to a newer version of prometheus/common.

---------

Signed-off-by: Vishal Choudhary <vishal.choudhary@nirmata.com>