feat: Collect triggers before processing UR for generate policies
The main advantage of this approach is to avoid sending list requests once for each trigger, even if they are all the same type; instead, we list once per (type,namespace) and re-use that list for all triggers of the same type. In situations with many triggers of one type, this means significantly lower load on the control plane.

Signed-off-by: Tomas Aschan <tomasl@spotify.com>