feat(engine): support platform selection in imageRegistry context
Adds a new optional 'platform' field to the imageRegistry context in policies. This allows users to specify the desired platform (e.g., "linux/arm64") when fetching image metadata from a multi-platform image index.

When the platform is specified, it is passed to the underlying
go-containerregistry library, resolving failures when looking up images that do not have a manifest for the default (linux/amd64) platform.

Fixes #13591

Signed-off-by: Rishi Chirchi <rishiraj.chirchi@gmail.com>