Add paths filter for conformance tests
This avoids running the tests for irrelevant changes, saving time and resources.
The dorny/paths-filer action is used to detect if the relevant files are touched.
Jobs then include this as a needed dependency and run based on the output of it.
The tricky part is the "aggregate" jobs (conformance-required-success
and conformance-required-failure). They need to now succeed/fail even if some
dependencies were skipped because no relevant files were touched.
We cannot use success() because that would fail if a job was skipped.
Instead we check if any of them failed explicitly (skipped is then allowed).

Signed-off-by: Lennart Jern <lennart.jern@est.tech>