Support other key algorithms than RSA
This adds support for more key algorithms when using externally provided certificates.
Kyverno will still generate only RSA keys but can now parse and validate other formats.
As verification, this commit enables the ECDSA chainsaw test that was previously failing
because of the lack of support in Kyverno.

Unit tests are also added for privateKeyToPem and pemToPrivateKey.
The main change to the code is the switch from rsa.PrivateKey to crypto.PrivateKey
in all relevant functions.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>