feat: Add --tlsKeyAlgorithm and support more algorithms
This adds support for other algorithms than RSA for Kyverno managed certificates.
It is now possible to choose between RSA (default), ECDSA and Ed25519.
The flag can be set through helm values admissionController.tlsKeyAlgorithm
and cleanupController.tlsKeyAlgorithm.

The certificate tests are refactored so that cert-manager and the certificates are handled through
the run-tests action directly. This reduces duplication and makes them more similar to other tests.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>